requirements.txt
bcrypt==4.0.1
Bcrypt-Flask==1.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
future==0.18.2
gevent==22.10.2
greenlet==2.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.3
psycopg2==2.9.5
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
requests-file==1.5.1
six==1.16.0
tldextract==3.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.0
